backend/requirement.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
SQLAlchemy>=1.4
python-dotenv>=0.21.0
psycopg2-binary>=2.9
PyMySQL>=1.0
